requirements.txt
elasticsearch
pyyaml
easydict
selenium
bs4